Move (almost) everything back to `Falcon::Environments`.
