+++ lib/falcon/command/redirect.rb
@@ -3,7 +3,7 @@
 # Released under the MIT License.
 # Copyright, 2020-2023, by Samuel Williams.
 
-require_relative '../service/redirect'
+require_relative '../environments/redirect'
 require_relative 'paths'
 
 require 'samovar'
@@ -30,7 +30,7 @@ module Falcon
 			include Paths
 			
 			def environment(**options)
-				Async::Service::Environment.new(Falcon::Service::Redirect::Environment).with(
+				Async::Service::Environment.new(Falcon::Environments::Redirect).with(
 					root: Dir.pwd,
 					verbose: self.parent&.verbose?,
 					url: @options[:bind],
@@ -44,7 +44,7 @@ module Falcon
 				hosts = {}
 				
 				environments.each do |environment|
-					next unless environment.implements?(Falcon::Service::Application::Environment)
+					next unless environment.implements?(Falcon::Environments::Application)
 					evaluator = environment.evaluator
 					hosts[evaluator.authority] = evaluator
 				end
