+++ lib/falcon/environments/rack.rb
@@ -3,13 +3,13 @@
 # Released under the MIT License.
 # Copyright, 2019-2023, by Samuel Williams.
 
-require_relative '../service/application'
+require_relative 'application'
 require_relative 'rackup'
 
 module Falcon
 	module Environments
 		module Rack
-			include Service::Application::Environment
+			include Application
 			include Rackup
 		end
